refactor(integration): update namespace to `SharedInfra` for better code organization
- Renamed `Infra` namespace to `SharedInfra` across v2 integration test files.
- Updated references in `SampleTest`, `TestFixture`, and `DynamoContainerFixture`.
- Improved logical grouping of shared infrastructure components.
