+++ tests/EntityFrameworkCore.DynamoDb.IntegrationTests.V2/SimpleTable/Infra/TestFixture.cs
@@ -1,5 +1,5 @@
 using Amazon.DynamoDBv2;
-using EntityFrameworkCore.DynamoDb.IntegrationTests.V2.Infra;
+using EntityFrameworkCore.DynamoDb.IntegrationTests.V2.SharedInfra;
 
 namespace EntityFrameworkCore.DynamoDb.IntegrationTests.V2.SimpleTable;
